requirements.txt
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argcomplete==3.6.3
asyncpg==0.30.0
certifi==2025.11.12
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
commitizen==4.10.0
decli==0.6.3
Deprecated==1.3.1
distlib==0.4.0
exceptiongroup==1.3.0
fastapi==0.121.2
filelock==3.19.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.11
importlib_metadata==8.6.1
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
nodeenv==1.9.1
packaging==25.0
passlib[bcrypt]==1.7.4
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
questionary==2.1.1
ruff==0.14.5
sendgrid==6.11.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
termcolor==3.1.0
tomli==2.3.0
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wrapt==2.0.1
zipp==3.23.0
